% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wacc.R
\name{wacc}
\alias{wacc}
\title{wacc}
\usage{
wacc(
  TC,
  coutstanding_shares,
  cstock_price,
  cstock_dividend,
  cstock_growth,
  cstock_beta,
  poutstanding_shares,
  pstock_price,
  pstock_beta,
  bvalue_debt,
  rf,
  cmrp,
  pmrp,
  coupon_rate
)
}
\arguments{
\item{TC}{}

\item{coutstanding_shares}{}

\item{cstock_price}{}

\item{cstock_dividend}{}

\item{cstock_growth}{}

\item{cstock_beta}{}

\item{poutstanding_shares}{}

\item{pstock_price}{}

\item{pstock_beta}{}

\item{bvalue_debt}{}

\item{rf}{}

\item{cmrp}{}

\item{pmrp}{}

\item{coupon_rate}{}
}
\value{

}
\description{
wacc
}
\examples{
wacc(
#corporate tax rate
TC = .22,
#common stock shares/price/dividend/growth/beta (yahoo finance)
coutstanding_shares = 346300407,
cstock_price = 108.12,
cstock_dividend = 4,
cstock_growth = .0444,
cstock_beta = .49,
#preferred stock shares/price/beta (yahoo finance)
poutstanding_shares = 346300407,
pstock_price = 108.12,
pstock_beta = .49,
#book value of debt
bvalue_debt = 35973686275,
#risk free rate (tYcharts.com/indicators/3_month_t_bill)
rf = .0232,
#market risk premium (www.market-risk-premia.com)
cmrp = .05,
pmrp = .05,
#market rate of new debt (coupon rate)
coupon_rate = .0395
)
}
